Improve consumable item sorting
Most items are sorted by their category and tier, keeping the random
order in which they appear in the `items` array from affecting the
totals box sorting. With consumable items, however, there are many
categories that fit into the same class, but have wildly different
functions. As such, this takes advantage of the pre-existing tier
sorting and extends it to apply to consumable items to better arrange
them in the totals box.

Closes atomizer/muledump#75

http://pastebin.com/gFWRiPDk