Sort consumable items by name rather than by ID
Because items change IDs fairly frequently when swaths of items are
reassigned positions in the game's XML definitions, or when new items
are introduced, sorting them based on their *current* ID is a fairly
fragile strategy which is in no way future-proof.

Instead, given that the item names and their associated tiers are known
and well-defined, we can instead sort using the item name as the
switch case.

This addresses:
https://github.com/atomizer/muledump/pull/80#issuecomment-142095938